Update BaseGradeBook.cs
Change return type from int to double

+++ GradeBook/GradeBooks/BaseGradeBook.cs
@@ -114,7 +114,7 @@ namespace GradeBook.GradeBooks
             }
         }
 
-        public virtual int GetGPA(char letterGrade, StudentType studentType)
+        public virtual double GetGPA(char letterGrade, StudentType studentType)
         {
             int gpa_equivalent = 0;
